changed package name and group id to linked-planet
