+++ src/main/java/com/linkedplanet/plugin/jira/insightobjectgraph/impl/package-info.java
@@ -1,0 +1,24 @@
+/**
+ * Contains Java classes we were forced to use instead of Kotlin in some scenarios.
+ */
+package com.linkedplanet.plugin.jira.insightobjectgraph.impl;
+
+/*-
+ * #%L
+ * insight-object-graph
+ * %%
+ * Copyright (C) 2018 - 2020 The Plugin Authors
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
